refactor(architecture): extract test scaffold and report into core
Move test command orchestration into core/extension/test/:
- scaffold.rs: scaffold workflow orchestration + output types
- report.rs: unified TestCommandOutput envelope + builders

commands/test.rs reduced from 725 to 387 lines (216 non-test).
The run() function is now pure CLI glue: parse args, delegate
to core workflows, shape output via report builders.
