+++ src/commands/test.rs
@@ -1,16 +1,11 @@
 use clap::Args;
-use serde::Serialize;
 use std::path::PathBuf;
 
-use homeboy::component::Component;
 use homeboy::extension::test as extension_test;
 use homeboy::extension::test::{
-    auto_fix_test_drift, detect_test_drift, CoverageOutput, DriftReport, TestAnalysis,
-    TestRunWorkflowArgs, TestScopeOutput, TestSummaryOutput,
+    auto_fix_test_drift, detect_test_drift, report, run_scaffold_workflow, TestCommandOutput,
+    TestRunWorkflowArgs,
 };
-use homeboy::refactor::AppliedRefactor;
-use homeboy::scaffold::ScaffoldConfig;
-use homeboy::extension::test::{TestBaselineComparison, TestCounts};
 
 use super::utils::args::{BaselineArgs, HiddenJsonArgs, PositionalComponentArgs, SettingArgs};
 use super::{CmdResult, GlobalArgs};
@@ -86,63 +81,6 @@ pub struct TestArgs {
     json_summary: bool,
 }
 
-#[derive(Serialize)]
-pub struct TestOutput {
-    status: String,
-    component: String,
-    exit_code: i32,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    test_counts: Option<TestCounts>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    coverage: Option<CoverageOutput>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    baseline_comparison: Option<TestBaselineComparison>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    analysis: Option<TestAnalysis>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    autofix: Option<AppliedRefactor>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    hints: Option<Vec<String>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    drift: Option<DriftReport>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    scaffold: Option<ScaffoldOutput>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    auto_fix_drift: Option<AutoFixDriftOutput>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    test_scope: Option<TestScopeOutput>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    summary: Option<TestSummaryOutput>,
-}
-
-#[derive(Serialize)]
-pub struct ScaffoldOutput {
-    results: Vec<ScaffoldFileOutput>,
-    total_stubs: usize,
-    total_written: usize,
-    total_skipped: usize,
-}
-
-#[derive(Serialize)]
-pub struct ScaffoldFileOutput {
-    source_file: String,
-    test_file: String,
-    stub_count: usize,
-    written: bool,
-    skipped: bool,
-}
-
-#[derive(Serialize)]
-pub struct AutoFixDriftOutput {
-    since: String,
-    auto_fixable_changes: usize,
-    generated_rules: usize,
-    replacements: usize,
-    files_modified: usize,
-    written: bool,
-    rerun_recommended: bool,
-}
-
 /// Filter out homeboy-owned flags from trailing args before passing to extension scripts.
 ///
 /// Clap's `trailing_var_arg = true` + `allow_hyphen_values = true` captures all arguments
@@ -216,28 +154,33 @@ fn filter_homeboy_flags(args: &[String]) -> Vec<String> {
     filtered
 }
 
-pub fn run(args: TestArgs, _global: &GlobalArgs) -> CmdResult<TestOutput> {
+pub fn run(args: TestArgs, _global: &GlobalArgs) -> CmdResult<TestCommandOutput> {
     let source_path = args.comp.source_path()?;
     let component = args.comp.load()?;
 
-    // Scaffold mode — generate test stubs without running tests
+    // Scaffold mode — delegate to core scaffold workflow
     if args.scaffold || args.scaffold_file.is_some() {
-        return run_scaffold(
+        let result = run_scaffold_workflow(
             args.comp.id(),
             &component,
             args.scaffold_file.as_deref(),
             args.write,
-        );
+        )?;
+        return Ok(report::from_scaffold_workflow(result.component, result.output));
     }
 
-    // Drift detection mode — skip running tests, analyze git changes instead
+    // Drift detection mode — delegate to core drift workflows
     if args.drift {
         if args.fix {
-            return run_auto_fix_drift(args.comp.id(), &component, &args.since, args.write, true);
+            let result =
+                auto_fix_test_drift(args.comp.id(), &component, &args.since, args.write, true)?;
+            return Ok(report::from_auto_fix_drift_workflow(result));
         }
-        return run_drift(args.comp.id(), &component, &args.since);
+        let result = detect_test_drift(args.comp.id(), &component, &args.since)?;
+        return Ok(report::from_drift_workflow(result));
     }
 
+    // Main test workflow — delegate to core
     let passthrough_args = filter_homeboy_flags(&args.args);
     let workflow = extension_test::run_main_test_workflow(
         &component,
@@ -261,293 +204,13 @@ pub fn run(args: TestArgs, _global: &GlobalArgs) -> CmdResult<TestOutput> {
         },
     )?;
 
-    Ok((
-        TestOutput {
-            status: workflow.status,
-            component: workflow.component,
-            exit_code: workflow.exit_code,
-            test_counts: workflow.test_counts,
-            coverage: workflow.coverage,
-            baseline_comparison: workflow.baseline_comparison,
-            analysis: workflow.analysis,
-            autofix: workflow.autofix,
-            hints: workflow.hints,
-            drift: None,
-            scaffold: None,
-            auto_fix_drift: None,
-            test_scope: workflow.test_scope,
-            summary: workflow.summary,
-        },
-        workflow.exit_code,
-    ))
-}
-
-/// Auto-fix test drift by generating transform rules from production changes.
-///
-/// This mode does NOT run tests. It inspects git changes since `since`, generates
-/// find/replace transform rules for auto-fixable drift types, and applies them to
-/// test files. Triggered by `homeboy test --drift --fix`.
-/// Use with `--write` to persist changes; default is dry-run.
-fn run_auto_fix_drift(
-    component_id: &str,
-    component: &Component,
-    since: &str,
-    write: bool,
-    include_report: bool,
-) -> CmdResult<TestOutput> {
-    let result = auto_fix_test_drift(component_id, component, since, write, include_report)?;
-
-    Ok((
-        TestOutput {
-            status: if result.output.replacements > 0 || !result.hints.is_empty() {
-                if write { "fixed" } else { "planned" }.to_string()
-            } else {
-                "passed".to_string()
-            },
-            component: result.component,
-            exit_code: 0,
-            test_counts: None,
-            coverage: None,
-            baseline_comparison: None,
-            analysis: None,
-            autofix: None,
-            hints: Some(result.hints),
-            drift: result.report,
-            scaffold: None,
-            auto_fix_drift: Some(AutoFixDriftOutput {
-                since: result.output.since,
-                auto_fixable_changes: result.output.auto_fixable_changes,
-                generated_rules: result.output.generated_rules,
-                replacements: result.output.replacements,
-                files_modified: result.output.files_modified,
-                written: result.output.written,
-                rerun_recommended: result.output.rerun_recommended,
-            }),
-            test_scope: None,
-            summary: None,
-        },
-        0,
-    ))
-}
-
-/// Run drift detection without running tests.
-fn run_drift(component_id: &str, component: &Component, since: &str) -> CmdResult<TestOutput> {
-    let result = detect_test_drift(component_id, component, since)?;
-
-    Ok((
-        TestOutput {
-            status: "drift".to_string(),
-            component: result.component,
-            exit_code: result.exit_code,
-            test_counts: None,
-            coverage: None,
-            baseline_comparison: None,
-            analysis: None,
-            autofix: None,
-            hints: None,
-            drift: Some(result.report),
-            scaffold: None,
-            auto_fix_drift: None,
-            test_scope: None,
-            summary: None,
-        },
-        result.exit_code,
-    ))
-}
-
-/// Run scaffold mode — generate test stubs from source files.
-fn run_scaffold(
-    component_id: &str,
-    component: &Component,
-    scaffold_file: Option<&str>,
-    write: bool,
-) -> CmdResult<TestOutput> {
-    let source_path = {
-        let expanded = shellexpand::tilde(&component.local_path);
-        std::path::PathBuf::from(expanded.as_ref())
-    };
-
-    // Auto-detect language
-    let config = if source_path.join("Cargo.toml").exists() {
-        ScaffoldConfig::rust()
-    } else {
-        ScaffoldConfig::php()
-    };
-
-    let mode_label = if write { "write" } else { "dry-run" };
-
-    if let Some(file) = scaffold_file {
-        // Single file mode
-        let file_path = source_path.join(file);
-        homeboy::log_status!(
-            "scaffold",
-            "Scaffolding tests for {} ({})",
-            file,
-            mode_label
-        );
-
-        let result = homeboy::scaffold::scaffold_file(&file_path, &source_path, &config, write)?;
-
-        if result.skipped {
-            homeboy::log_status!(
-                "scaffold",
-                "Skipped — test file already exists: {}",
-                result.test_file
-            );
-        } else if result.stub_count == 0 {
-            homeboy::log_status!("scaffold", "No public methods found in {}", file);
-        } else {
-            homeboy::log_status!(
-                "scaffold",
-                "Generated {} test stub{} → {}{}",
-                result.stub_count,
-                if result.stub_count == 1 { "" } else { "s" },
-                result.test_file,
-                if write { " (written)" } else { " (dry-run)" }
-            );
-
-            if !write {
-                // Show preview
-                eprintln!("---");
-                for line in result.content.lines().take(40) {
-                    eprintln!("{}", line);
-                }
-                if result.content.lines().count() > 40 {
-                    eprintln!("... ({} more lines)", result.content.lines().count() - 40);
-                }
-                eprintln!("---");
-            }
-        }
-
-        let scaffold_output = ScaffoldOutput {
-            results: vec![ScaffoldFileOutput {
-                source_file: result.source_file.clone(),
-                test_file: result.test_file.clone(),
-                stub_count: result.stub_count,
-                written: result.written,
-                skipped: result.skipped,
-            }],
-            total_stubs: result.stub_count,
-            total_written: if result.written { 1 } else { 0 },
-            total_skipped: if result.skipped { 1 } else { 0 },
-        };
-
-        Ok((
-            TestOutput {
-                status: "scaffold".to_string(),
-                component: component_id.to_string(),
-                exit_code: 0,
-                test_counts: None,
-                coverage: None,
-                baseline_comparison: None,
-                analysis: None,
-                autofix: None,
-                hints: None,
-                drift: None,
-                scaffold: Some(scaffold_output),
-                auto_fix_drift: None,
-                test_scope: None,
-                summary: None,
-            },
-            0,
-        ))
-    } else {
-        // Batch mode — scaffold all untested files
-        homeboy::log_status!(
-            "scaffold",
-            "Scanning {} for untested {} files ({})",
-            component_id,
-            config.language,
-            mode_label
-        );
-
-        let batch = homeboy::scaffold::scaffold_untested(&source_path, &config, write)?;
-
-        let files_needing_tests = batch
-            .results
-            .iter()
-            .filter(|r| !r.skipped && r.stub_count > 0)
-            .count();
-        let already_tested = batch.total_skipped;
-
-        homeboy::log_status!(
-            "scaffold",
-            "{} file{} need tests, {} already have tests",
-            files_needing_tests,
-            if files_needing_tests == 1 { "" } else { "s" },
-            already_tested
-        );
-
-        if write {
-            homeboy::log_status!(
-                "scaffold",
-                "Wrote {} test file{} with {} total stubs",
-                batch.total_written,
-                if batch.total_written == 1 { "" } else { "s" },
-                batch.total_stubs
-            );
-        } else if files_needing_tests > 0 {
-            // Show summary of what would be created
-            for result in &batch.results {
-                if !result.skipped && result.stub_count > 0 {
-                    homeboy::log_status!(
-                        "  new",
-                        "{} → {} ({} stubs)",
-                        result.source_file,
-                        result.test_file,
-                        result.stub_count
-                    );
-                }
-            }
-            homeboy::log_status!(
-                "hint",
-                "Run with --write to create test files: homeboy test {} --scaffold --write",
-                component_id
-            );
-        }
-
-        let scaffold_output = ScaffoldOutput {
-            results: batch
-                .results
-                .iter()
-                .map(|r| ScaffoldFileOutput {
-                    source_file: r.source_file.clone(),
-                    test_file: r.test_file.clone(),
-                    stub_count: r.stub_count,
-                    written: r.written,
-                    skipped: r.skipped,
-                })
-                .collect(),
-            total_stubs: batch.total_stubs,
-            total_written: batch.total_written,
-            total_skipped: batch.total_skipped,
-        };
-
-        Ok((
-            TestOutput {
-                status: "scaffold".to_string(),
-                component: component_id.to_string(),
-                exit_code: 0,
-                test_counts: None,
-                coverage: None,
-                baseline_comparison: None,
-                analysis: None,
-                autofix: None,
-                hints: None,
-                drift: None,
-                scaffold: Some(scaffold_output),
-                auto_fix_drift: None,
-                test_scope: None,
-                summary: None,
-            },
-            0,
-        ))
-    }
+    Ok(report::from_main_workflow(workflow))
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use homeboy::component::Component;
     use homeboy::refactor::test_refactor_request;
     use homeboy::refactor::TestSourceOptions;
